+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Cache entity registry index; rebuild only on entity changes
+- SVG Annotation Caching
 
 ## [0.1.10] - 2026-01-27
 ### Added